app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Timur Logachov <ifk.webdev@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
mongo = "app.mongo:main"
scrapping = "app.scrapping:main"

[tool.poetry.dependencies]
python = "^3.11"
pymongo = "^4.7.2"
requests = "^2.31.0"
bs4 = "^0.0.2"
lxml = "^5.2.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
